server/requirements.txt
fastapi==0.52.0
langchain==0.2.6
langchain_community==0.2.6
langchain_google_genai==1.0.6
pydantic==1.10.12
python-dotenv==1.0.1
supabase==2.5.1
uvicorn==0.30.1
